inputtestfiles

+++ src/Road.cpp
@@ -97,7 +97,7 @@ int Road::getNextBusStop(unsigned int currentPosition) {
     for (std::set<unsigned int>::iterator it = fBusStops.begin(); it != fBusStops.end(); ++it) {
         if (currentPosition < *it)return *it;
     }
-    for (int i = 0; i < fConnections.size(); ++i) {
+    for (unsigned int i = 0; i < (unsigned)fConnections.size(); ++i) {
         return fConnections[i]->getNextBusStop(0);
     }
     return -1;
